+++ src/core/mod.rs
@@ -1,4 +1,5 @@
 mod connection_ready_state;
+mod direction;
 mod element_maybe_signal;
 mod elements_maybe_signal;
 mod maybe_rw_signal;
@@ -9,6 +10,7 @@ mod ssr_safe_method;
 mod storage;
 
 pub use connection_ready_state::*;
+pub use direction::*;
 pub use element_maybe_signal::*;
 pub use elements_maybe_signal::*;
 pub use maybe_rw_signal::*;